feat: Hack a way to make two-paragraph page intros.
This is pretty hacky and probably needs some refining, but it works for 
now. In the long run, I think what we want to do here is use Markdown to 
parse the summaries, so we can format them more granularly (ie, bolds 
would be good!) and so we can use smart quotes, but I couldn't figure 
out how to do this, and I just needed a way to make line breaks, so this 
will have to do for now.

See also #60.