feat: Implement down-arrow navigation between portfolio panels.
Note: this is currently relying on manually passing the title of the 
next panel, which is obviously a bit fragile and prone to breakage. It 
would be better if we could programmatically determine the hash of the 
next panel and automatically include it, but I don't know how to make 
this happen!