fix: Add a manual canonical URL so Gatsby's happy.
This is a total hack, since we don't actually need a canonical URL
for any of these pages, but we want to be sure that we can override them
if/when we need to.

So, for now, this does the trick, but it would be better replaced with
smarter logic in the SEO component.