fix: Add a canonical URL to about page.
This isn't strictly necessary except Gatsby needs at least one page to
have canonical frontmatter in order to make that data available in
GraphQL.

So, we're going to add a completely unnecessary canonical URL to this one page,
on the off chance we might some day want to use this feature.

Ghost Ship will need to do the same, for obvious reasons!